# Improve performance on multiple versions