--- a/cmd/xl-storage-format-utils.go
+++ b/cmd/xl-storage-format-utils.go
@@ -150,20 +150,26 @@ func getXLDiskLoc(diskID string) (poolIdx, setIdx, diskIdx int) {
 	return -1, -1, -1
 }
 
-// hashDeterministicString will return a deterministic (weak) hash for the map values.
+// hashDeterministicString will return a deterministic hash for the map values.
+// Trivial collisions are avoided, but this is by no means a strong hash.
 func hashDeterministicString(m map[string]string) uint64 {
+	// Seed (random)
 	var crc = uint64(0xc2b40bbac11a7295)
+	// Xor each value to make order independent
 	for k, v := range m {
-		crc = crc ^ (xxh3.HashString(k) + xxh3.HashString(v))
+		// Separate key and value with an individual xor with a random number.
+		// Add values of each, so they cannot be trivially collided.
+		crc = crc ^ ((xxh3.HashString(k) ^ 0x4ee3bbaf7ab2506b) + (xxh3.HashString(v) ^ 0x8da4c8da66194257))
 	}
 	return crc
 }
 
 // hashDeterministicBytes will return a deterministic (weak) hash for the map values.
+// Trivial collisions are avoided, but this is by no means a strong hash.
 func hashDeterministicBytes(m map[string][]byte) uint64 {
 	var crc = uint64(0x1bbc7e1dde654743)
 	for k, v := range m {
-		crc = crc ^ (xxh3.HashString(k) + xxh3.Hash(v))
+		crc = crc ^ ((xxh3.HashString(k) ^ 0x4ee3bbaf7ab2506b) + (xxh3.Hash(v) ^ 0x8da4c8da66194257))
 	}
 	return crc
 }