--- a/cmd/xl-storage-format-utils.go
+++ b/cmd/xl-storage-format-utils.go
@@ -19,7 +19,6 @@ package cmd
 
 import (
 	"fmt"
-	"sort"
 
 	jsoniter "github.com/json-iterator/go"
 	"github.com/minio/minio/internal/logger"