--- a/cmd/xl-storage-format-v1.go
+++ b/cmd/xl-storage-format-v1.go
@@ -23,8 +23,10 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/cespare/xxhash/v2"
 	jsoniter "github.com/json-iterator/go"
 	"github.com/minio/minio/internal/logger"
+	"github.com/tinylib/msgp/msgp"
 )
 
 // XL constants.
@@ -201,6 +203,30 @@ func (m *xlMetaV1Object) ToFileInfo(volume, path string) (FileInfo, error) {
 		DataDir:   m.DataDir,
 	}
 	return fi, nil
+}
+
+// Signature will return a signature that is expected to be the same across all disks.
+func (m *xlMetaV1Object) Signature() [4]byte {
+	// Shallow copy
+	c := *m
+	// Zero unimportant fields
+	c.Erasure.Index = 0
+	c.Minio.Release = ""
+	x := xxhash.New()
+
+	// Use JSON for maps, since it sorts keys.
+	json := jsoniter.ConfigCompatibleWithStandardLibrary
+	enc := json.NewEncoder(x)
+	enc.Encode(c.Meta)
+	c.Meta = nil
+
+	if err := msgp.Encode(x, &c); err != nil {
+		return signatureErr
+	}
+	var tmp [4]byte
+	// Signature is lower 32 bits
+	copy(tmp[:], x.Sum(nil))
+	return tmp
 }
 
 // XL metadata constants.