… cmd/testdata/xl.meta-v1.2.zst
… cmd/xl-storage-format-utils_test.go
… cmd/xl-storage-format-v2_gen.go
… cmd/xl-storage-format-v2_gen_test.go
… cmd/xl-storage-format-v2_string.go
… cmd/xl-storage-format-v2_test.go
… cmd/xl-storage-format_test.go
… cmd/xl-storage-free-version.go
… cmd/xl-storage-free-version_test.go
… cmd/xl-storage-meta-inline.go
… cmd/xl-storage.go
… go.mod
… go.sum