## Description

Existing:

``` Go
type xlMetaV2 struct {
    Versions []xlMetaV2Version `json:"Versions" msg:"Versions"`
}
```

Serialized as regular MessagePack.

Change to add a shallow header and each version separately:

```Go
//msgp:tuple xlMetaV2VersionHeader
type xlMetaV2VersionHeader struct {
	VersionID [16]byte
	ModTime   int64
	Type      VersionType
	Flags     xlFlags
}
```

Serialize as streaming MessagePack, format:

```
int(headerVersion)
int(xlmetaVersion)
int(nVersions)
for each version {
    binary blob, xlMetaV2VersionHeader, serialized
    binary blob, xlMetaV2Version, serialized.
}
```

xlMetaV2VersionHeader is <= 30 bytes serialized. Deserialized struct can easily be reused and does not contain pointers, so efficient as a slice (single alloc)

This allows quickly parsing everything as slices of bytes (no copy).

Versions are always *saved* sorted by modtime, newest *first*. No more need to sort on load.

* Allows checking if a version exists.
* Allows reading single version without unmarshal all.
* Allows reading latest version of type without unmarshal all.
* Allows reading latest version without unmarshal of all.
* Allows checking if latest is deletemarker by reading first entry.
* Allows adding/updating/deleting a version with only header deserialization.
* Reduces allocations on conversion to FileInfo(s).

## How to test this PR?

Tiering and Replication should probably be checked.

## Types of changes
- [x] Optimization (provides speedup with no functional changes)

## Checklist:
- [ ] Documentation updated (xl-meta executable)
- [x] Unit tests added/updated
- [ ] Optimize loading existing
- [ ] Optimize saving size.
- [ ] Benchmarks.