@klauspost 2021-11-04
@harshavardhana Here is to show the approximate scaling changes for realistic operations:

`size-1` is one version, `size-10` 10 versions, etc..
```
λ benchcmp before.txt after.txt
benchmark                                                       old ns/op     new ns/op     delta
BenchmarkXlMetaV2Shallow/size-1/UpdateObjectVersion-32          2247          3056          +36.00%
BenchmarkXlMetaV2Shallow/size-1/DeleteVersion-32                1226          1951          +59.14%
BenchmarkXlMetaV2Shallow/size-1/AddVersion-32                   4116          2273          -44.78%
BenchmarkXlMetaV2Shallow/size-1/ToFileInfo-32                   7512          7342          -2.26%
BenchmarkXlMetaV2Shallow/size-10/UpdateObjectVersion-32         16593         4634          -72.07%
BenchmarkXlMetaV2Shallow/size-10/AddVersion-32                  18481         3361          -81.81%
BenchmarkXlMetaV2Shallow/size-1000/UpdateObjectVersion-32       1632360       164222        -89.94%
BenchmarkXlMetaV2Shallow/size-1000/AddVersion-32                1641188       119245        -92.73%
BenchmarkXlMetaV2Shallow/size-100000/UpdateObjectVersion-32     179432900     27262307      -84.81%
BenchmarkXlMetaV2Shallow/size-100000/DeleteVersion-32           176398333     52705835      -70.12%
BenchmarkXlMetaV2Shallow/size-100000/AddVersion-32              175290350     21089523      -87.97%
BenchmarkXlMetaV2Shallow/size-100000/ToFileInfo-32              200999380     21906535      -89.10%

benchmark                                                       old MB/s     new MB/s     speedup
BenchmarkXlMetaV2Shallow/size-1/UpdateObjectVersion-32          0.44         0.33         0.75x
BenchmarkXlMetaV2Shallow/size-1/DeleteVersion-32                0.82         0.51         0.62x
BenchmarkXlMetaV2Shallow/size-1/AddVersion-32                   0.24         0.44         1.83x
BenchmarkXlMetaV2Shallow/size-1/ToFileInfo-32                   0.13         0.14         1.08x
BenchmarkXlMetaV2Shallow/size-10/UpdateObjectVersion-32         0.60         2.16         3.60x
BenchmarkXlMetaV2Shallow/size-10/AddVersion-32                  0.54         2.98         5.52x
BenchmarkXlMetaV2Shallow/size-1000/UpdateObjectVersion-32       0.61         6.09         9.98x
BenchmarkXlMetaV2Shallow/size-1000/AddVersion-32                0.61         8.39         13.75x
BenchmarkXlMetaV2Shallow/size-100000/UpdateObjectVersion-32     0.56         3.67         6.55x
BenchmarkXlMetaV2Shallow/size-100000/DeleteVersion-32           0.57         1.90         3.33x
BenchmarkXlMetaV2Shallow/size-100000/AddVersion-32              0.57         4.74         8.32x
BenchmarkXlMetaV2Shallow/size-100000/ToFileInfo-32              0.50         4.56         9.12x

benchmark                                                       old allocs     new allocs     delta
BenchmarkXlMetaV2Shallow/size-1/UpdateObjectVersion-32          13             26             +100.00%
BenchmarkXlMetaV2Shallow/size-1/DeleteVersion-32                12             24             +100.00%
BenchmarkXlMetaV2Shallow/size-1/AddVersion-32                   26             13             -50.00%
BenchmarkXlMetaV2Shallow/size-1/ToFileInfo-32                   50             60             +20.00%
BenchmarkXlMetaV2Shallow/size-10/UpdateObjectVersion-32         103            29             -71.84%
BenchmarkXlMetaV2Shallow/size-10/AddVersion-32                  116            13             -88.79%
BenchmarkXlMetaV2Shallow/size-1000/UpdateObjectVersion-32       10003          30             -99.70%
BenchmarkXlMetaV2Shallow/size-1000/AddVersion-32                10016          13             -99.87%
BenchmarkXlMetaV2Shallow/size-100000/UpdateObjectVersion-32     1000003        30             -100.00%
BenchmarkXlMetaV2Shallow/size-100000/DeleteVersion-32           1000002        100023         -90.00%
BenchmarkXlMetaV2Shallow/size-100000/AddVersion-32              1000016        13             -100.00%
BenchmarkXlMetaV2Shallow/size-100000/ToFileInfo-32              1000043        60             -99.99%

benchmark                                                       old bytes     new bytes     delta
BenchmarkXlMetaV2Shallow/size-1/UpdateObjectVersion-32          264           2027          +667.80%
BenchmarkXlMetaV2Shallow/size-1/DeleteVersion-32                304           1426          +369.08%
BenchmarkXlMetaV2Shallow/size-1/AddVersion-32                   1400          881           -37.07%
BenchmarkXlMetaV2Shallow/size-1/ToFileInfo-32                   1736          2798          +61.18%
BenchmarkXlMetaV2Shallow/size-10/UpdateObjectVersion-32         2136          2175          +1.83%
BenchmarkXlMetaV2Shallow/size-10/AddVersion-32                  3272          881           -73.07%
BenchmarkXlMetaV2Shallow/size-1000/UpdateObjectVersion-32       208056        2189          -98.95%
BenchmarkXlMetaV2Shallow/size-1000/AddVersion-32                209192        881           -99.58%
BenchmarkXlMetaV2Shallow/size-100000/UpdateObjectVersion-32     20800056      2662          -99.99%
BenchmarkXlMetaV2Shallow/size-100000/DeleteVersion-32           20800098      1602240       -92.30%
BenchmarkXlMetaV2Shallow/size-100000/AddVersion-32              20801194      1180          -99.99%
BenchmarkXlMetaV2Shallow/size-100000/ToFileInfo-32              24004673      2950          -99.99% 
```

I will get the single object performance to equal the current before making it ready for review.

There were some strange random failures with the previous code, so some results are missing. I will see if I can track it down.

Also "ToFileInfo" has a specialized function, approx 2x the speed of the one used above:
```
BenchmarkXlMetaV2Shallow/size-1/ToFileInfoNew-32                  171352              7112 ns/op           0.14 MB/s        2768 B/op         59 allocs/op
BenchmarkXlMetaV2Shallow/size-10/ToFileInfoNew-32                 154939              7747 ns/op           1.29 MB/s        2768 B/op         59 allocs/op
BenchmarkXlMetaV2Shallow/size-1000/ToFileInfoNew-32                19075             62666 ns/op          15.96 MB/s        2768 B/op         59 allocs/op
BenchmarkXlMetaV2Shallow/size-100000/ToFileInfoNew-32                100          11185109 ns/op           8.94 MB/s        2768 B/op         59 allocs/op
```